Split files and cleanup a bit
