+++ src/main/java/com/justifiedsolutions/openpdf/pdf/internal/JSPDFWriter.java
@@ -8,8 +8,6 @@ package com.justifiedsolutions.openpdf.pdf.internal;
 
 import com.justifiedsolutions.openpdf.pdf.Chapter;
 import com.justifiedsolutions.openpdf.pdf.Document;
-import com.justifiedsolutions.openpdf.pdf.Footer;
-import com.justifiedsolutions.openpdf.pdf.Header;
 import com.justifiedsolutions.openpdf.pdf.Margin;
 import com.justifiedsolutions.openpdf.pdf.Metadata;
 import com.justifiedsolutions.openpdf.pdf.PageSize;
@@ -22,10 +20,7 @@ import com.justifiedsolutions.openpdf.text.Element;
 import com.justifiedsolutions.openpdf.text.Meta;
 import com.justifiedsolutions.openpdf.text.Rectangle;
 import com.justifiedsolutions.openpdf.text.RectangleReadOnly;
-import com.justifiedsolutions.openpdf.text.pdf.ColumnText;
-import com.justifiedsolutions.openpdf.text.pdf.PdfContentByte;
 import com.justifiedsolutions.openpdf.text.pdf.PdfPTable;
-import com.justifiedsolutions.openpdf.text.pdf.PdfPageEventHelper;
 import com.justifiedsolutions.openpdf.text.pdf.PdfWriter;
 import java.io.OutputStream;
 import java.util.Map;
@@ -99,52 +94,5 @@ public class JSPDFWriter {
             result = PdfPTable.getInstance((Table) content);
         }
         return result;
-    }
-
-    private static class HeaderFooterHelper extends PdfPageEventHelper {
-
-        private final Header header;
-        private final Footer footer;
-
-        public HeaderFooterHelper(Header header, Footer footer) {
-            this.header = header;
-            this.footer = footer;
-        }
-
-        @Override
-        public void onStartPage(PdfWriter writer,
-                com.justifiedsolutions.openpdf.text.Document document) {
-            if (header != null && header.isValidForPageNumber(document.getPageNumber())) {
-                writeHeaderFooter(writer, document, header.getParagraph(document.getPageNumber()),
-                        document.top() + 27);
-            }
-        }
-
-        @Override
-        public void onEndPage(PdfWriter writer,
-                com.justifiedsolutions.openpdf.text.Document document) {
-            if (footer != null && footer.isValidForPageNumber(document.getPageNumber())) {
-                writeHeaderFooter(writer, document, footer.getParagraph(document.getPageNumber()),
-                        document.bottom() - 27);
-            }
-        }
-
-        private void writeHeaderFooter(PdfWriter writer, com.justifiedsolutions.openpdf.text.Document document,
-                Paragraph modelParagraph, float y) {
-            com.justifiedsolutions.openpdf.text.Paragraph paragraph = com.justifiedsolutions.openpdf.text.Paragraph
-                    .getInstance(modelParagraph);
-            float x;
-            int alignment = paragraph.getAlignment();
-            if (alignment == Element.ALIGN_LEFT) {
-                x = document.left();
-            } else if (alignment == Element.ALIGN_RIGHT) {
-                x = document.right();
-            } else {
-                x = document.getPageSize().getRight() / 2.0f;
-                alignment = Element.ALIGN_CENTER;
-            }
-            final PdfContentByte cb = writer.getDirectContent();
-            ColumnText.showTextAligned(cb, alignment, paragraph, x, y, 0);
-        }
     }
 }
